feat(Client-Waiting-Sheet): Client waiting with map actions and api handling

+++ androidApp/src/main/java/uz/ildam/technologies/yalla/android/ui/screens/cancel_reason/CancelReasonViewModel.kt
@@ -43,7 +43,7 @@ class CancelReasonViewModel(
             uiState.value.selectedReason?.apply {
                 when (cancelReasonUseCase(
                     orderId = AppPreferences.lastOrderId,
-                    reasonId = id,
+                    reasonId = id.toInt(),
                     reasonComment = name
                 )) {
                     is Result.Error -> _actionState.emit(CancelReasonActionState.Error)
