+++ androidApp/src/main/java/uz/ildam/technologies/yalla/android/ui/screens/history/HistoryIntent.kt
@@ -2,5 +2,5 @@ package uz.ildam.technologies.yalla.android.ui.screens.history
 
 sealed interface HistoryIntent {
     data object OnNavigateBack : HistoryIntent
-    data class OnHistoryItemClick(val id: Int) : HistoryIntent
+    data class OnHistoryItemClick(val id: Long) : HistoryIntent
 }
