+++ androidApp/src/main/java/uz/ildam/technologies/yalla/android/ui/screens/map/MapActionState.kt
@@ -2,7 +2,7 @@ package uz.ildam.technologies.yalla.android.ui.screens.map
 
 sealed interface MapActionState {
     data object LoadingAddressId : MapActionState
-    data class AddressIdLoaded(val id: Int) : MapActionState
+    data class AddressIdLoaded(val id: Long) : MapActionState
     data object LoadingAddressName : MapActionState
     data class AddressNameLoaded(val name: String?) : MapActionState
     data object LoadingPolygon : MapActionState
